リファクタ一通り
