+++ .gitignore
@@ -4,3 +4,5 @@ bundle.js
 bundle.js.LICENSE.txt
 example.db
 shortcut.db
+template.db
+clipboard.db
